requirements.txt
Flask==3.0.3 #3.1.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Login==0.6.3
Werkzeug==3.1.3
python-dotenv==1.0.0 #1.1.1
pytesseract==0.3.13
Pillow==11.1.0
pypdf==5.6.0
pdf2image==1.17.0
email_validator==2.2.0
psycopg2-binary==2.9.10
boto3==1.38.41
pytest==8.4.1